--- Controller/User/Registration.ts
@@ -1,697 +1,0 @@
-import User from "../../model/UserSchema";
-import { Request, Response } from "express";
-import bcrypt from "bcrypt";
-import jwt, { JwtPayload } from "jsonwebtoken";
-import mongoose from "mongoose";
-import { CustomError } from "../../Utils/errorHandler";
-
-
-
-//Registration
-const RegistrationUser = async (req: Request, res: Response): Promise<void> => {
-  const {
-    email,
-    password,
-    firstName,
-    lastName,
-    education,
-    location,
-    jobTitles,
-    jobLocations,
-  } = req.body;
-
-  const emailRegex = /\S+@\S+\.\S+/;
-  if (!emailRegex.test(email)) {
-    throw new CustomError('Invalid email format',400)
-  }
-
-  const existingUser = await User.findOne({ email });
-  if (existingUser) {
-    throw new CustomError('User already exists',400)
-  }
-
-  const hashedPassword = await bcrypt.hash(password, 10);
-
-  const user = new User({
-    email,
-    password: hashedPassword,
-    firstName,
-    lastName,
-    education: education || [],
-    location,
-    jobTitle: jobTitles,
-    jobLocation: jobLocations,
-  });
-  await user.save();
-
-  const token = jwt.sign(
-    { id: user._id, email: user.email },
-    process.env.USER_SECRETKEY!,
-    { expiresIn: "1d" }
-  );
-  const refreshToken = jwt.sign(
-    { id: user._id, email: user.email },
-    process.env.USER_SECRETKEY!,
-    { expiresIn: "7d" }
-  );
-
-  res.cookie("token", token, {
-    httpOnly: true,
-    secure: true,
-    sameSite: "lax",
-    maxAge: 24 * 60 * 60 * 1000,
-  });
-
-  res.cookie("refreshToken", refreshToken, {
-    httpOnly: true,
-    secure: true,
-    sameSite: "lax",
-    maxAge: 7 * 24 * 60 * 60 * 1000,
-  });
-
-  res
-    .status(200)
-    .json({ status: true, message: "Registration successful", user });
-};
-
-////////////////////// LOGIN API ////////////////////////
-
-const login = async (req: Request, res: Response): Promise<void> => {
-  const { email, password } = req.body;
-  const logeduser = await User.findOne({ email });
-  if (!logeduser) {
-    throw new CustomError("email id is wrong",404);
-   
-  }
-  const verfyuser = await bcrypt.compare(password, logeduser.password);
-  if (!verfyuser) {
-    throw new CustomError("password is wrong",404);
-    
-  }
-  const currentDate = new Date();
-  if (logeduser.role === "premium" && logeduser.subscriptionEndDate) {
-    if (logeduser.subscriptionEndDate < currentDate) {
-      logeduser.role = "user";
-      logeduser.subscriptionStartDate = null;
-      logeduser.subscriptionEndDate = null;
-      await logeduser.save();
-    }
-  }
-  if (verfyuser) {
-    const token = jwt.sign(
-      {
-        id: logeduser._id,
-        email: logeduser.email,
-        isBlocked: logeduser.isBlocked,
-      },
-      process.env.USER_SECRETKEY!,
-      { expiresIn: "1d" }
-    );
-    res.cookie("token", token, {
-      httpOnly: true,
-      secure: true,
-      sameSite: "lax",
-      maxAge: 24 * 60 * 60 * 1000,
-    });
-    const refreshToken = jwt.sign(
-      { id: logeduser._id, email: logeduser.email },
-      process.env.USER_SECRETKEY!,
-      { expiresIn: "7d" }
-    );
-    res.cookie("refreshToken", refreshToken, {
-      httpOnly: true,
-      secure: true,
-      sameSite: "lax",
-      maxAge: 7 * 24 * 60 * 60 * 1000,
-    });
-  }
-
-  if (logeduser.role === "premium" && logeduser.subscriptionEndDate) {
-    const subscriptionEndDate = logeduser.subscriptionEndDate
-      ? new Date(logeduser.subscriptionEndDate)
-      : null;
-
-    if (subscriptionEndDate && !isNaN(subscriptionEndDate.getTime())) {
-      const currentDate = new Date();
-
-      const startOfDay = (date: Date) => new Date(date.setHours(0, 0, 0, 0));
-
-      const normalizedEndDate = startOfDay(subscriptionEndDate);
-      const normalizedCurrentDate = startOfDay(currentDate);
-
-      const differenceInTime =
-        normalizedEndDate.getTime() - normalizedCurrentDate.getTime();
-
-      const remainingValidityDays = Math.floor(
-        differenceInTime / (1000 * 60 * 60 * 24)
-      );
-
-      if (remainingValidityDays > 0) {
-        const payload = {
-          userId: logeduser._id,
-          email: logeduser.email,
-          role: logeduser.role,
-          remainingValidityDays,
-        };
-
-        const secretKey = process.env.USER_SECRETKEY!;
-
-        const subscriptionToken = jwt.sign(payload, secretKey, {
-          expiresIn: `${remainingValidityDays}d`,
-        });
-
-        res.cookie("subscriptionToken", subscriptionToken, {
-          httpOnly: false,
-          secure: process.env.NODE_ENV === "production",
-          sameSite: "lax",
-          path: "/",
-          maxAge: remainingValidityDays * 24 * 60 * 60 * 1000,
-        });
-      } else {
-        throw new CustomError(
-          "Your premium membership has expired. Please renew to continue enjoying premium benefits.",
-          403
-        )
-      }
-    } else {
-      console.error("Invalid subscription end date");
-    }
-  }
-
-  res
-    .status(200)
-    .json({ status: true, message: "Login successful", logeduser });
-};
-
-//////////////////////////////  LOGOUT //////////////////
-
-const logout = async (req: Request, res: Response): Promise<void> => {
-  const userId = req.user?.id;
-
-  if (userId) {
-    const user = await User.findById(userId);
-
-    if (user && user.role === "premium" && user.subscriptionEndDate) {
-      const currentDate = new Date();
-
-      if (user.subscriptionEndDate < currentDate) {
-        user.role = "user";
-        user.subscriptionStartDate = null;
-        user.subscriptionEndDate = null;
-        await user.save();
-      }
-    }
-  }
-
-  res.clearCookie("token", {
-    httpOnly: true,
-    secure: false,
-    sameSite: "lax",
-  });
-  res.clearCookie("refreshToken", {
-    httpOnly: true,
-    secure: false,
-    sameSite: "lax",
-  });
-
-  res.clearCookie("subscriptionToken", {
-    httpOnly: true,
-    secure: false,
-    sameSite: "lax",
-  });
-
-  res.status(200).json({ status: true, message: "Logout successfully" });
-};
-
-////////////////////// GOOGLE AUTH LOGIN /////////////////////
-
-const googleauthlogin = async (req: Request, res: Response) => {
-  const { email, name, image } = req.body;
-
-  
-  if(!name && !email){
-    throw new CustomError("name or email is missing",404);
-   
-  }
-  const finduser = await User.findOne({ email });
-
-  if (finduser) {
-    const token = jwt.sign(
-      {
-        id: finduser._id,
-        email: finduser.email,
-        isBlocked: finduser.isBlocked,
-      },
-      process.env.USER_SECRETKEY!,
-      { expiresIn: "1d" }
-    );
-
-    const currentDate = new Date();
-    if (finduser.role === "premium" && finduser.subscriptionEndDate) {
-      if (finduser.subscriptionEndDate < currentDate) {
-        finduser.role = "user";
-        finduser.subscriptionStartDate = null;
-        finduser.subscriptionEndDate = null;
-        await finduser.save();
-      }
-    }
-
-    if (finduser.role === "premium" && finduser.subscriptionEndDate) {
-      const subscriptionEndDate = finduser.subscriptionEndDate
-        ? new Date(finduser.subscriptionEndDate)
-        : null;
-
-      if (subscriptionEndDate && !isNaN(subscriptionEndDate.getTime())) {
-        const currentDate = new Date();
-
-        const startOfDay = (date: Date) => new Date(date.setHours(0, 0, 0, 0));
-
-        const normalizedEndDate = startOfDay(subscriptionEndDate);
-        const normalizedCurrentDate = startOfDay(currentDate);
-
-        const differenceInTime =
-          normalizedEndDate.getTime() - normalizedCurrentDate.getTime();
-
-        const remainingValidityDays = Math.floor(
-          differenceInTime / (1000 * 60 * 60 * 24)
-        );
-
-        if (remainingValidityDays > 0) {
-          const payload = {
-            userId: finduser._id,
-            email: finduser.email,
-            role: finduser.role,
-            remainingValidityDays,
-          };
-
-          const secretKey = process.env.USER_SECRETKEY!;
-
-          const subscriptionToken = jwt.sign(payload, secretKey, {
-            expiresIn: `${remainingValidityDays}d`,
-          });
-
-          res.cookie("subscriptionToken", subscriptionToken, {
-            httpOnly: false,
-            secure: process.env.NODE_ENV === "production",
-            sameSite: "lax",
-            path: "/",
-            maxAge: remainingValidityDays * 24 * 60 * 60 * 1000,
-          });
-        } else {
-          throw new CustomError(
-            "Your premium membership has expired. Please renew to continue enjoying premium benefits.",
-            404
-          )
-        }
-      } else {
-      }
-    }
-    res.cookie("token", token, {
-      httpOnly: true,
-      secure: true,
-      sameSite: "lax",
-      maxAge: 24 * 60 * 60 * 1000,
-    });
-    const refreshToken = jwt.sign(
-      {
-        id: finduser._id,
-        email: finduser.email,
-        isBlocked: finduser.isBlocked,
-      },
-      process.env.USER_SECRETKEY!,
-      { expiresIn: "1d" }
-    );
-    res.cookie("refreshToken", refreshToken, {
-      httpOnly: true,
-      secure: true,
-      sameSite: "lax",
-      maxAge: 24 * 60 * 60 * 1000,
-    });
-    res.status(200).json({
-      status: true,
-      message: "google auth Login successful",
-      finduser,
-    });
-    return;
-  } else {
-    const user = await new User({
-      email,
-      firstName: name,
-      profileImage: image,
-    });
-    const savegoogleauth = await user.save();
-    res.status(200).json({
-      status: true,
-      message: "google auth registration and Login successful",
-      savegoogleauth,
-    });
-  }
-};
-
-const findCurrentUserDetails = async (
-  req: Request,
-  res: Response
-): Promise<void> => {
-  const userId = req.user?.id;
-  if (!userId) {
-    throw new CustomError("User ID is missing",404);
-  }
-
-  const currentUserDetails = await User.findById(userId).select("-password");
-
-  if (!currentUserDetails) {
-    throw new CustomError("User not found",404);
-    
-  }
-
-  res.status(200).json({ success: true, currentUserDetails });
-};
-
-export const getPeopleYouMightKnow = async (
-  req: Request,
-  res: Response
-): Promise<void> => {
-  // const userId = req.user?.id;
-  // if (!userId || !mongoose.Types.ObjectId.isValid(userId)) {
-  //   res.status(400).json({ error: "Valid User ID is required" });
-  //   return;
-  // }
-  // const user = await User.findById(userId).populate("connecting");
-  // if (!user) {
-  //   res.status(404).json({ error: "User not found" });
-  //   return;
-  // }
-  // const potentialConnections = await User.find({
-  //   _id: {
-  //     $ne: userId,
-  //     $nin: user.connecting?.map((conn) => conn._id) || []
-  //   }
-  // }).populate("connecting");
-  // const suggestedPeople = potentialConnections.filter((person) => {
-  //   // Find mutual connections
-  //   const mutualConnections = user.connecting?.filter((conn) =>
-  //     person.connecting?.some((pConn) => pConn._id.equals(conn._id))
-  //   );
-  //   return mutualConnections.length > 0;
-  // });
-  // res.status(200).json({ suggestedPeople });
-};
-
-export const updateUserProfile = async (req: Request, res: Response): Promise<void> => {
-
-    const userId = req.user?.id; // Assuming `req.user` is set from authentication middleware
-    if (!userId || !mongoose.Types.ObjectId.isValid(userId)) {
-      throw new CustomError("Valid User ID is required",404);
-      
-    }
-
-  // Find user
-  const user = await User.findById(userId);
-  if (!user) {
-    throw new CustomError("User not found",404);
-  }
-
-    
-    const {
-      firstName,
-      lastName,
-      email,
-      phoneNumber,
-      dateOfBirth,
-      location,
-      skills,
-      jobTitle,
-      jobLocation,
-      about,
-      education,
-      projects,
-      profileImage,
-      banner
-    } = req.body;
-
-    const updateData: { [key: string]: any } = {
-      ...(firstName && { firstName }),
-      ...(lastName && { lastName }),
-      ...(email && { email }),
-      ...(phoneNumber && { phoneNumber }),
-      ...(dateOfBirth && { dateOfBirth }),
-      ...(location && { location }),
-      ...(skills && { skills }),
-      ...(jobTitle && { jobTitle }),
-      ...(jobLocation && { jobLocation }),
-      ...(about && { about }),
-      ...(education && { education }),
-      ...(projects && { projects }),
-      ...(profileImage && { profileImage }),
-      ...(banner && { banner }),
-
-    };
-
-  const updatedUser = await User.findByIdAndUpdate(userId, updateData, {
-    new: true,
-  });
-
-    if (!updatedUser) {
-      throw new CustomError("User not found",404);
-      
-    }
-
-    res.status(200).json({status:true, message: "Profile updated successfully", user: updatedUser });
-  
-
-
-};
-
-
-
-export const uploadResume = async (req: Request, res: Response): Promise<void> => {
-
-    const files = req.files as {
-      resume?: Express.Multer.File[];
-      video?: Express.Multer.File[];
-    };
-    const pdfFile = files?.resume ? files.resume[0] : null;
-    const videoFile = files?.video ? files.video[0] : null;
-
-
-    if (!pdfFile && !videoFile) {
-      throw new CustomError("No files uploaded",400)
-
-    }
-
-    const userId = req.user?.id;
-    const user = await User.findById(userId);
-
-    if (!user) {
-      throw new CustomError("User not found",400)
-
-    }
-
-    if (!user.resumePDF) user.resumePDF = [];
-    if (!user.resumeVideo) user.resumeVideo = [];
-
-    if (pdfFile) {
-      const existingActivePDF = user.resumePDF.some((pdf) => !pdf.isDeleted);
-      
-      if (!existingActivePDF) {
-        user.resumePDF.push({
-          fileUrl: pdfFile.path,
-          fileName: pdfFile.originalname,
-          uploadedAt: new Date(),
-          isDeleted: false,
-        });
-      } else {
-        throw new CustomError("A resume PDF already exists.",400)
-    
-      }
-    }
-
-    if (videoFile) {
-      const existingActiveVideo = user.resumeVideo.some(
-        (video) => !video.isDeleted
-      );
-      if (!existingActiveVideo) {
-        user.resumeVideo.push({
-          fileUrl: videoFile.path,
-          fileName: videoFile.originalname,
-          uploadedAt: new Date(),
-          isDeleted: false,
-        });
-      } else {
-        throw new CustomError("A resume video already exists.",400)
-   
-      }
-    }
-
-    await user.save();
-
-    res.status(200).json({
-      success: true,
-      message: "Resume uploaded successfully",
-      user: {
-        ...user.toObject(),
-        resumePDF: user.resumePDF.filter((pdf) => !pdf.isDeleted),
-        resumeVideo: user.resumeVideo.filter((video) => !video.isDeleted),
-      },
-    });
-
-};
-
-export const getUploadedFiles = async (
-  req: Request,
-  res: Response
-): Promise<void> => {
-
-    const userId = req.user?.id;
-    console.log("user", userId);
-
-    if (!userId) {
-      throw new CustomError("Unauthorized",401)
-
-    }
-
-    const user = await User.findById(userId);
-
-    if (!user) {
-      throw new CustomError("User not found",401)
- 
-    }
-
-    const activeResumePDFs = user.resumePDF?.filter((pdf) => !pdf.isDeleted);
-    const activeResumeVideos = user.resumeVideo?.filter(
-      (video) => !video.isDeleted
-    );
-
-    res.status(200).json({
-      success: true,
-      message: "Uploaded files retrieved successfully",
-      uploadedFiles: {
-        resumePDFs: activeResumePDFs,
-        resumeVideos: activeResumeVideos,
-      },
-    });
-
-};
-
-const removeResumeFile = async (req: Request, res: Response): Promise<void> => {
-  const userId: string | undefined = req.user?.id;
-  const fileType: "resume" | "introductionVideo" | undefined = req.query
-    .fileType as "resume" | "introductionVideo";
-
-
-  if (!userId) {
-    throw new CustomError("Unauthorized",401)
-
-  }
-
-  if (
-    !fileType ||
-    (fileType !== "resume" && fileType !== "introductionVideo")
-  ) {
-    throw new CustomError("Invalid file type. Use 'pdf' or 'video'.",400)
-
-  }
-
-  const user = await User.findById(userId);
-  if (!user) {
-    throw new CustomError("User not found",401)
-
-  }
-
-  let updatedFiles = [];
-  if (fileType === "resume" && Array.isArray(user.resumePDF)) {
-    user.resumePDF.forEach((resume) => {
-      resume.isDeleted = true;
-    });
-
-    updatedFiles = user.resumePDF.filter((resume) => !resume.isDeleted);
-  } else if (
-    fileType === "introductionVideo" &&
-    Array.isArray(user.resumeVideo)
-  ) {
-    user.resumeVideo.forEach((resume) => {
-      resume.isDeleted = true;
-    });
-
-    updatedFiles = user.resumeVideo.filter((resume) => !resume.isDeleted);
-  } else {
-    throw new CustomError(`No ${fileType} files found to mark as deleted`,404)
-
-  }
-
-    await user.save();
-    res.status(200).json({
-      success: true,
-      message: `${fileType} files marked as deleted.`,
-      files: updatedFiles,
-    });
-
-};
-
-
-//Email check
-
-const AllUsersEmailCheck = async (req: Request, res: Response) => {
-  const { email } = req.query;
-
-  const user = await User.findOne({ email });
-
-  if (user) {
-    res.json({ exists: true });
-    return;
-  } else {
-    res.json({ exists: false });
-    return;
-  }
-};
-
-//All users
-const AllUsers = async (req: Request, res: Response) => {
-  const users = await User.find();
-  const length = users.length;
-  if (!users) {
-    throw new CustomError("users not found",404)
-  }
-
-  res
-    .status(200)
-    .json({ status: true, message: "all users detailes", users, length });
-  return;
-};
-
-////////////////// ALL USER PROFILE /////////////////
-
-const spacificuserdetails = async (
-  req: Request,
-  res: Response
-): Promise<void> => {
-  const userid = req.params.id;
-
-  if(!userid){
-    res.status(404).json({status:false,message:"cannot find id"})
-    return
-  }
-  const finduserprofile = await User.findOne({_id:userid,isDeleted:false,isBlocked:false}).populate('connecting.connectionID')
-
-  
-  if(!finduserprofile){
-    res.status(404).json({status:false,message:"cannot find all profile"})
-    return
-  }
-
-  res
-    .status(200)
-    .json({ status: true, message: "All profile finded", finduserprofile });
-};
-
-export {
-  RegistrationUser,
-  login,
-  logout,
-  findCurrentUserDetails,
-  googleauthlogin,
-  AllUsersEmailCheck,
-  AllUsers,
-  spacificuserdetails,
-  removeResumeFile,
-};
