+++ Controller/commentController/user.ts
@@ -1,0 +1,142 @@
+import { Request, Response } from "express";
+import { Comment } from "../../model/CommentSchema";
+import { Post } from "../../model/PostSchema";
+import { CustomError } from "../../Utils/errorHandler";
+
+// Get all comments
+export const getAllComments = async (
+  req: Request,
+  res: Response
+): Promise<void> => {
+  const comments = await Comment.find().populate("user");
+  const totalComments = await Comment.countDocuments();
+  res.status(200).json({
+    status: true,
+    message: "Got all the comments",
+    comments,
+    totalComments,
+  });
+};
+
+// Comment on a Post
+export const addCommentToPost = async (
+  req: Request,
+  res: Response
+): Promise<void> => {
+  const { postId, comment } = req.body;
+
+  if (!comment || comment.trim() === "") {
+    throw new CustomError("Comment cannot be empty", 400);
+  }
+
+  const post = await Post.findById(postId);
+  if (!post) {
+    throw new CustomError("Post not found", 404);
+  }
+
+  const newComment = new Comment({
+    user: req.user?.id,
+    comment,
+  });
+
+  await newComment.save();
+
+  if (!post.comments) {
+    post.comments = [];
+  }
+  post.comments.push(newComment.id);
+
+  await post.save();
+  const populatedComment = await Comment.findById(newComment._id)
+    .populate("user", "firstName lastName email profileImage")
+    .exec();
+
+  res.status(201).json({
+    status: true,
+    message: "Comment added successfully",
+    comment: populatedComment,
+  });
+};
+
+// update a comment by id
+export const editComment = async (
+  req: Request,
+  res: Response
+): Promise<void> => {
+  const { newComment } = req.body;
+  const { commentId } = req.params;
+
+  if (!newComment || newComment.trim() === "") {
+    throw new CustomError("Comment cannot be empty", 400);
+  }
+
+  const comment = await Comment.findById(commentId);
+  if (!comment) {
+    throw new CustomError("Comment not found", 404);
+  }
+
+  if (comment.user.toString() !== req.user?.id) {
+    throw new CustomError(
+      "Unauthorized: You can only edit your own comments",
+      403
+    );
+  }
+  comment.comment = newComment;
+  await comment.save();
+
+  res.status(200).json({
+    status: true,
+    message: "Comment updated successfully",
+    comment,
+  });
+  return;
+};
+
+// Delete a Comment
+export const deleteComment = async (
+  req: Request,
+  res: Response
+): Promise<void> => {
+  const { commentId } = req.params;
+
+  const comment = await Comment.findById(commentId);
+  if (!comment) {
+    throw new CustomError("Comment not found", 404);
+  }
+
+  if (comment.user.toString() !== req.user?.id) {
+    throw new CustomError(
+      "Unauthorized: You can only delete your own comments",
+      403
+    );
+  }
+
+  comment.isDeleted = true;
+  await comment.save();
+
+  res.status(200).json({
+    status: true,
+    message: "Comment marked as deleted successfully",
+    comment,
+  });
+};
+
+// get a comment by ID
+export const getCommentById = async (
+  req: Request,
+  res: Response
+): Promise<void> => {
+  const comment = await Comment.findById(req.params.id);
+
+  if (!comment) {
+    throw new CustomError("No post found containing this comment", 404);
+  }
+
+  const commented = comment.comment;
+
+  if (!commented) {
+    throw new CustomError("Comment not found", 404);
+  }
+  res.status(200).json({ status: true, message: "Comment found", comment });
+  return;
+};
