+++ index.ts
@@ -3,15 +3,14 @@ import dotenv from "dotenv";
 import { connectMongodb } from "./config/dbconections";
 import cors from 'cors';
 import cookieParser from "cookie-parser";
-import { router } from "./Routes/user/registration";
-import { rating } from "./Routes/Rating/rating";
-import postRouter from "./Routes/PostRoute/post";
-import { paymentRouter } from "./Routes/PaymentRoute/PaymentRoutes";
-import{Searchrouter} from'./Routes/SearchRoute/search';
-
-import { connectionroute } from "./Routes/user/connection";
-import { adminRoutes } from "./Routes/Admin/AdminRoutes";
+import { userRouter } from "./routes/userRoutes";
+import { ratingRouter } from "./routes/ratingRoutes";
+import { paymentRouter } from "./routes/PaymentRoutes";
+import{searchRouter} from'./routes/searchRoutes';
+import { adminRouter } from "./routes/AdminRoutes";
 import errorHandler from "./middleware/customClassMiddleware";
+import postRouter from "./routes/postRoutes";
+import { connectionRouter } from "./routes/connectionRoutes";
 
 dotenv.config();
 const app = express();
@@ -24,13 +23,13 @@ app.use(cors({
     origin: 'http://localhost:3000',   
     credentials: true,
   }));
-  app.use("/api/user",router)
-  app.use("/api/rating",rating) 
+  app.use("/api/user",userRouter)
+  app.use("/api/rating",ratingRouter) 
   app.use("/api/post",postRouter)
   app.use("/api/payment",paymentRouter)
-  app.use("/api/user",Searchrouter)
-  app.use("/api/connecting",connectionroute)
-  app.use("/api/admin",adminRoutes)
+  app.use("/api/user",searchRouter)
+  app.use("/api/connecting",connectionRouter)
+  app.use("/api/admin",adminRouter)
 
   app.use(errorHandler);
